requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17
jinja2==3.1.4
aiofiles==24.1.0

# Database
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==43.0.3

# Task queue
celery==5.4.0
redis==5.2.0
kombu==5.4.2

# HTTP client
httpx==0.28.0
tenacity==9.0.0

# Config
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1

# AI / APIs
anthropic==0.40.0
openai==1.57.0

# Email
aiosmtplib==3.0.2

# Utils
python-dateutil==2.9.0.post0
pytz==2024.2
pillow==11.0.0
boto3==1.35.81

# Monitoring (optional)
prometheus-fastapi-instrumentator==7.0.0
